# Improve immutablility of MMap