After:

```
[info] Benchmark                                             Mode  Cnt       Score      Error   Units
[info] GenericBench.decodeFoo                               thrpt   20  343880.808 ± 4494.298   ops/s
[info] GenericBench.decodeFoo:·gc.alloc.rate                thrpt   20    5596.513 ±   72.122  MB/sec
[info] GenericBench.decodeFoo:·gc.alloc.rate.norm           thrpt   20   18784.001 ±   14.255    B/op
[info] GenericBench.decodeFoo:·gc.churn.G1_Eden_Space       thrpt   20    5674.959 ±   82.305  MB/sec
[info] GenericBench.decodeFoo:·gc.churn.G1_Eden_Space.norm  thrpt   20   19047.610 ±  160.700    B/op
[info] GenericBench.decodeFoo:·gc.churn.G1_Old_Gen          thrpt   20       0.002 ±    0.001  MB/sec
[info] GenericBench.decodeFoo:·gc.churn.G1_Old_Gen.norm     thrpt   20       0.008 ±    0.002    B/op
[info] GenericBench.decodeFoo:·gc.count                     thrpt   20     611.000             counts
[info] GenericBench.decodeFoo:·gc.time                      thrpt   20     793.000                 ms
[info] GenericBench.encodeFoo                               thrpt   20  434349.884 ± 5455.212   ops/s
[info] GenericBench.encodeFoo:·gc.alloc.rate                thrpt   20    3843.980 ±   46.776  MB/sec
[info] GenericBench.encodeFoo:·gc.alloc.rate.norm           thrpt   20   10216.001 ±   14.255    B/op
[info] GenericBench.encodeFoo:·gc.churn.G1_Eden_Space       thrpt   20    3872.510 ±   94.855  MB/sec
[info] GenericBench.encodeFoo:·gc.churn.G1_Eden_Space.norm  thrpt   20   10290.516 ±  163.473    B/op
[info] GenericBench.encodeFoo:·gc.churn.G1_Old_Gen          thrpt   20       0.001 ±    0.001  MB/sec
[info] GenericBench.encodeFoo:·gc.churn.G1_Old_Gen.norm     thrpt   20       0.003 ±    0.001    B/op
[info] GenericBench.encodeFoo:·gc.count                     thrpt   20     417.000             counts
[info] GenericBench.encodeFoo:·gc.time                      thrpt   20     560.000                 ms
```

Before:

```
[info] Benchmark                                             Mode  Cnt       Score       Error   Units
[info] GenericBench.decodeFoo                               thrpt   20  475918.467 ±  9135.225   ops/s
[info] GenericBench.decodeFoo:·gc.alloc.rate                thrpt   20    7556.937 ±   133.773  MB/sec
[info] GenericBench.decodeFoo:·gc.alloc.rate.norm           thrpt   20   18328.001 ±    35.637    B/op
[info] GenericBench.decodeFoo:·gc.churn.G1_Eden_Space       thrpt   20    7661.238 ±   155.571  MB/sec
[info] GenericBench.decodeFoo:·gc.churn.G1_Eden_Space.norm  thrpt   20   18580.429 ±   142.582    B/op
[info] GenericBench.decodeFoo:·gc.churn.G1_Old_Gen          thrpt   20       0.004 ±     0.001  MB/sec
[info] GenericBench.decodeFoo:·gc.churn.G1_Old_Gen.norm     thrpt   20       0.010 ±     0.002    B/op
[info] GenericBench.decodeFoo:·gc.count                     thrpt   20     825.000              counts
[info] GenericBench.decodeFoo:·gc.time                      thrpt   20    1080.000                  ms
[info] GenericBench.encodeFoo                               thrpt   20  612437.225 ± 13027.842   ops/s
[info] GenericBench.encodeFoo:·gc.alloc.rate                thrpt   20    5190.974 ±   113.353  MB/sec
[info] GenericBench.encodeFoo:·gc.alloc.rate.norm           thrpt   20    9784.000 ±     7.127    B/op
[info] GenericBench.encodeFoo:·gc.churn.G1_Eden_Space       thrpt   20    5246.507 ±   137.911  MB/sec
[info] GenericBench.encodeFoo:·gc.churn.G1_Eden_Space.norm  thrpt   20    9888.165 ±   114.119    B/op
[info] GenericBench.encodeFoo:·gc.churn.G1_Old_Gen          thrpt   20       0.002 ±     0.001  MB/sec
[info] GenericBench.encodeFoo:·gc.churn.G1_Old_Gen.norm     thrpt   20       0.005 ±     0.001    B/op
[info] GenericBench.encodeFoo:·gc.count                     thrpt   20     565.000              counts
[info] GenericBench.encodeFoo:·gc.time                      thrpt   20     744.000                  ms
```